requirements.txt
alpaca-py>=0.8.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0
torch>=2.0.0
pandas-ta>=0.3.14b
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
loguru>=0.7.0
rich>=13.0.0
websockets>=11.0
aiohttp>=3.9.0
scipy>=1.11.0
scikit-learn>=1.3.0
plotext>=5.2.8
shimmy>=1.3.0
